Move grouped rendering to ruff_db
Summary
--

This is a pretty large refactor. The `grouped` output format depends on code
from the more general `TextEmitter`, so in addition to moving over the `grouped`
module itself, I also copied the necessary supporting code from
`ruff_linter::message::text` and `ruff_linter::message::diff`. The `text` module
also depended on the `ruff_linter::line_width` module, so I moved that to
`ruff_db` too. Similarly, I moved `Locator::ceil_char_boundary` to a free
function in `ruff_db`, along with a TODO to make it private once the `text`
module is moved over.

Test Plan
--

Existing tests ported to `ruff_db`
